fix: stack bounds
The primer and the ISA doc state for stack access the offset is computed
by taking the lowest 16 bits.
This commit creates a little trait to easily extract those from `U256`
and `u64`, as well as lowest 32 bits (for stack pointer, that is at most
u16::MAX+1 due to being the biggest stack address + 1).
Doing this fixes a panic when storing to the stack using big values.