Fix ret opcode
This wasn't breaking any test, but when doing reverts and ok, we were not clearing the registers and the context as specified in the spec. Since ret.panic, ret.revert and ret.ok have very similar behavior, I though it was proper to unify them in ret to make sure we ran the same common behaviour in all of them.